ai-rag-assistant/requirements.txt
sentence-transformers
numpy
streamlit
requests
python-dotenv